feat(runtime-c-api) Extract the `import` module.
